requirements.txt
pytest
pytest-html
airtest
allure-pytest
PyYAML
requests
Pillow
uiautomator2
packaging
